Nested classes. Create Picture (linked by foreign key with Shops)
-ShortCut (+o-). Start duplicate project from Major Key (Shops)
-Rename and refactor methods and attributes in Domain, service,
controller and repository.
--First in new Entity (linked with foreign key). Create attributes,
getters and setters, @ManyToOne (in this case) and instantiate the
foreign key. Refactor the constructor to fit needs.
--Start to refactor Request Methods by editing Service and controller,
request by request. The easiest order is POST,PUT(similar to
PUT),DELETE,GET, and finally GET ALL.
---Note about GET ALL, you have to link in the Repository the method to
find a particular object filtering first by ForeignKey. Spring does it
simnple (Check Picture CrudRepository class) but you have to be very
precise with the CamlCase and the statement findBy_____ Completing with
the NAME OF THE VARIABLE that is the primary key in the Superior related
Entity. (In this case is findByShopName since ShopName is the name of
the primary key in the Shop class)

TO DO:
-Check Date format problem in instantiation
-Reject new pictures if the shop capacity is reached
-Create Destroy all records functions